use gpg1 and let the user generate the key
aptly only supports gpg1, as gpg defaults to gpg2 at some point, we have
to call gpg1 explictly.

Additionally we probably want to use an existing gpg key instead of
creating a new one in most cases. Therefore we will just check if a key is
available and fail if there is none.

We still keep the config for key generation on the server for testing
purposes. Alongside that, the command for key generation is shown on the
command line in case of missing key.