add aptly_upload script for package upload via ssh
aptly_upload is intended to run as ForceCommand for a specific ssh-key.
Therefore it takes all parameters through stdin. It expects a version id
in the first line of the stream, json data on the second line followed by the
actual .deb package.

The json-data carries the following information:
	- package name
	- target os/distro (currently only debian/buster)

multiple targets may be defined.

aptly_upload will iterate through all given targets, add the package to
each of them and updates the repositories appropriately.